extends AbstractAIAssistant
