+++ bigtop-manager-ai/bigtop-manager-ai-openai/src/main/java/org/apache/bigtop/manager/ai/openai/OpenAIAssistant.java
@@ -22,7 +22,6 @@ import org.apache.bigtop.manager.ai.core.AbstractAIAssistant;
 import org.apache.bigtop.manager.ai.core.enums.PlatformType;
 import org.apache.bigtop.manager.ai.core.exception.AssistantConfigNotSetException;
 import org.apache.bigtop.manager.ai.core.factory.AIAssistant;
-import org.apache.bigtop.manager.ai.core.provider.AIAssistantConfigProvider;
 
 import dev.langchain4j.data.message.AiMessage;
 import dev.langchain4j.data.message.SystemMessage;
@@ -36,7 +35,6 @@ import dev.langchain4j.model.chat.StreamingChatLanguageModel;
 import dev.langchain4j.model.openai.OpenAiChatModel;
 import dev.langchain4j.model.openai.OpenAiStreamingChatModel;
 import dev.langchain4j.model.output.Response;
-import dev.langchain4j.store.memory.chat.ChatMemoryStore;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.FluxSink;
 
@@ -128,28 +126,7 @@ public class OpenAIAssistant extends AbstractAIAssistant {
         return new Builder();
     }
 
-    public static class Builder implements AIAssistant.Builder {
-        private Object id;
-
-        private ChatMemoryStore chatMemoryStore;
-        private AIAssistantConfigProvider configProvider;
-
-        public Builder() {}
-
-        public Builder withConfigProvider(AIAssistantConfigProvider configProvider) {
-            this.configProvider = configProvider;
-            return this;
-        }
-
-        public Builder id(Object id) {
-            this.id = id;
-            return this;
-        }
-
-        public Builder memoryStore(ChatMemoryStore chatMemoryStore) {
-            this.chatMemoryStore = chatMemoryStore;
-            return this;
-        }
+    public static class Builder extends AbstractAIAssistant.Builder {
 
         public AIAssistant build() {
             String model = ValidationUtils.ensureNotNull(configProvider.getModel(), "model");
